Move back to f64

+++ src/t_df.rs
@@ -3,7 +3,7 @@ use crate::sphere::angular_distance_and_bearing;
 use ndarray::{Array1, Array2, Array3, Zip};
 
 /// Physical constant: permeability of free space (µ0)
-const MU0: f32 = 1e-7;
+const MU0: f64 = 1e-7;
 
 /// Calculates the "Transfer Matrix" (T) for Divergence-Free Spherical Elementary Current Systems (SECS).
 ///
@@ -37,16 +37,16 @@ const MU0: f32 = 1e-7;
 /// * `secs_altitude` - Altitude above the Earth's surface at which poles are located in meters.
 ///
 /// # Returns
-/// `Array3<f32>` representing the transfer matrix T, with dimensions [nobs][3][nsec].
+/// `Array3<f64>` representing the transfer matrix T, with dimensions [nobs][3][nsec].
 pub fn t_df(
     obs_locs: &[GeographicalPoint],
-    obs_altitude: f32,
+    obs_altitude: f64,
     secs_locs: &[GeographicalPoint],
-    secs_altitude: f32,
-) -> Array3<f32> {
+    secs_altitude: f64,
+) -> Array3<f64> {
     let nobs = obs_locs.len();
     let nsec = secs_locs.len();
-    let mut t = Array3::<f32>::zeros((nobs, 3, nsec));
+    let mut t = Array3::<f64>::zeros((nobs, 3, nsec));
 
     let (theta, alpha) = angular_distance_and_bearing(&obs_locs, &secs_locs);
 
@@ -111,7 +111,7 @@ mod tests {
 
     #[test]
     fn test_t_df_few_points_under() {
-        let t: Array3<f32> = t_df(
+        let t: Array3<f64> = t_df(
             // Substract R_EARTH in order to have same results as python code as `t_df`
             // adds R_EARTH
             &[
@@ -128,7 +128,7 @@ mod tests {
         );
 
         // values generated from the python code
-        let expected: Array3<f32> = Array3::from_shape_vec(
+        let expected: Array3<f64> = Array3::from_shape_vec(
             (2, 3, 3),
             vec![
                 -3.67250861e-11,
@@ -163,7 +163,7 @@ mod tests {
 
     #[test]
     fn test_t_df_few_points_over() {
-        let t: Array3<f32> = t_df(
+        let t: Array3<f64> = t_df(
             // Substract R_EARTH in order to have same results as python code as `t_df`
             // adds R_EARTH
             &[
@@ -180,7 +180,7 @@ mod tests {
         );
 
         // values generated from the python code
-        let expected: Array3<f32> = Array3::from_shape_vec(
+        let expected: Array3<f64> = Array3::from_shape_vec(
             (2, 3, 3),
             vec![
                 1.248881305890098e-11,
